rafactor: email-user domain 구분
