+++ src/main/java/com/gdg/poppet/auth/domain/converter/AuthConverter.java
@@ -1,6 +1,6 @@
 package com.gdg.poppet.auth.domain.converter;
 
-import com.gdg.poppet.user.domain.enums.EmailPeriod;
+import com.gdg.poppet.email.domain.enums.EmailPeriod;
 import com.gdg.poppet.user.domain.enums.Gender;
 import com.gdg.poppet.user.domain.model.User;
